+++ models/__init__.py
@@ -6,6 +6,7 @@ from . import insight_task_dependency
 from . import insight_employee_shift
 from . import project_project
 from . import report_cost_report
+from . import report_gantt_report
 from . import project_task
 from . import project_milestone
 from . import tjp_parser